% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken.R
\name{broken}
\alias{broken}
\title{broken}
\usage{
broken(xgb.model, new_observation, opt = "interactions", baseline = 0)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{new_observation}{a new observation}

\item{opt}{= "interaction" or "single"}

\item{baseline}{}
}
\value{
an object of the broken class
}
\description{
Function \code{broken}
}
