% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDepth.R
\name{calculateDepth}
\alias{calculateDepth}
\title{calculateDepth}
\usage{
calculateDepth(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
This function returns list of trees with column \code{depth}.
}
