% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGain.R
\name{calculateGain}
\alias{calculateGain}
\title{caculateGain}
\usage{
calculateGain(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{calculateGain} returns list of trees with pairs of variable and their gain.
}
