% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateGainTable.R
\name{calculateGainTable}
\alias{calculateGainTable}
\title{Table of pairs' gains}
\usage{
calculateGainTable(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{calculateGainTable} returns table of gains of all variables' pair occures in the model.
}
