% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePairsGainTable.R
\name{calculatePairsGainTable}
\alias{calculatePairsGainTable}
\title{Table of pairs' gains}
\usage{
calculatePairsGainTable(xgb.model, data)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{
a data table
}
\description{
Table containing gains of all variables' pairs occur in the model.
}
\examples{

}
