% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPairs.R
\name{countPairs}
\alias{countPairs}
\title{Table of occurancess number}
\usage{
countPairs(xgb.model, data)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{

}
\description{
Table containing occurancess number of variables' pairs in the model.
}
\examples{

}
