% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRoots.R
\name{countRoots}
\alias{countRoots}
\title{Table with number of roots}
\usage{
countRoots(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{countRoots} counts how many times each variable is in the root of the tree and calculates the weighedRoot-number of occurrences in root weighed by Gain.
}
