% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EIX-package}
\alias{EIX-package}
\title{EIX package}
\description{
Structure mining from 'XGBoost' and 'LightGBM' models.
Key functionalities of this package cover: visualisation of tree-based ensembles models,
identification of interactions, measuring of variable importance,
measuring of interaction importance, explanation of single prediction
with break down plots (based on 'xgboostExplainer' and 'iBreakDown' packages).
To download the 'LightGBM' use the following link: <https://github.com/Microsoft/LightGBM>.
EIX' is a part of the 'DrWhy.AI' universe.
}
