% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairsPlot.R
\name{ggpairsPlot}
\alias{ggpairsPlot}
\title{ggpairsPlot}
\usage{
ggpairsPlot(xgb.model, data)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{
a ggplot object
}
\description{
ggpairsPlot
}
\examples{

}
