% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceInteractions.R
\name{importanceInteractions}
\alias{importanceInteractions}
\title{Importance table for interactions.}
\usage{
importanceInteractions(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{importanceInteractions} returns table with importance of interactions with the following measure:
\itemize{
\item sumGain
\item sumCover
\item mean5Gain
\item meanGain
\item meanCover
\item freqency
}
}
