% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancePlot.R
\name{importancePlot}
\alias{importancePlot}
\title{Importance plot}
\usage{
importancePlot(xgb.model, data, xlab = "sumCover", ylab = "sumGain",
  opt = "mixed", top = 10)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{xlab}{measure on the x-axis. Default "sumCover"}

\item{ylab}{measure on the y-axis. Default "sumGain"}

\item{opt}{if "single" then plot includes only single variable,
if "interactions", then only interactions
if "mixed", then both single variable and interactons.
Default "mixed".}

\item{top}{number of positions on the plot or NULL for all variable. Default 10.}
}
\value{
a ggplot object
}
\description{
The plot containing two chosen measures of variables' importance in the model.
}
\details{
Available measures:
\itemize{
\item "sumGain" - sum of Gain value in all nodes, in which given variable occurs
\item "sumCover" - sum of Cover value in all nodes, in which given variable occurs; for LightGBM models: number of observation, which pass through the node
\item "mean5Gain" - mean gain from 5 occurrences of given variable with the highest gain
\item "meanGain" - mean Gain value in all nodes, in which given variable occurs
\item "meanCover" - mean Cover value in all nodes, in which given variable occurs; for LightGBM models: mean number of observation, which pass through the node
\item "freqency" - number of occurrences in the nodes for given variable
}

Additionally for plots with single variables:
\itemize{
\item "meanDepth"  - mean depth weighted by gain
\item "counterRoot" - number of occurrences in the root
\item "weightedRoot" - mean number of occurrences in the root, which is weighted by gain
}
}
\examples{

}
