% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceRadarPlot.R
\name{importanceRadarPlot}
\alias{importanceRadarPlot}
\title{Importance radar plot}
\usage{
importanceRadarPlot(xgb.model, data, opt = "mixed", top = 10,
  trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{opt}{"single", "mixed","interactions". Default "mixed"}

\item{top}{Default 10}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Importance radar plot
}
