% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceSingleVariable.R
\name{importanceSingleVariable}
\alias{importanceSingleVariable}
\title{Impotrance table for single variable}
\usage{
importanceSingleVariable(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Impotrance table for single variable
}
