% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceTable.R
\name{importanceTable}
\alias{importanceTable}
\title{Importance table}
\usage{
importanceTable(xgb.model, data, option = "both")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{option}{if "variables" then table includes only single variable,
if "interactions", then only interactions
if "both", then both single variable and interactons.
Default "both".}
}
\value{
a data table
}
\description{
Importance table
}
\details{
Available measures:
\itemize{
\item "sumGain" - sum of Gain value in all nodes, in which given variable occurs
\item "sumCover" - sum of Cover value in all nodes, in which given variable occurs; for LightGBM models: number of observation, which pass through the node
\item "mean5Gain" - mean gain from 5 occurrences of given variable with the highest gain
\item "meanGain" - mean Gain value in all nodes, in which given variable occurs
\item "meanCover" - mean Cover value in all nodes, in which given variable occurs; for LightGBM models: mean number of observation, which pass through the node
\item "freqency" - number of occurrences in the nodes for given variable
}

Additionally for table with single variables:
\itemize{
\item "meanDepth"  - mean depth weighted by gain
\item "numberOfRoots" - number of occurrences in the root
\item "weightedRoot" - mean number of occurrences in the root, which is weighted by gain
}
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

imp <- importanceTable(xgb.model, sm, option = "both")
imp
plot(imp,  top = 10)

imp <- importanceTable(xgb.model, sm, option = "variables")
imp
plot(imp,  top = nrow(imp))

 imp <- importanceTable(xgb.model, sm, option = "interactions")
 imp
plot(imp,  top = 10)

 imp <- importanceTable(xgb.model, sm, option = "variables")
 imp
plot(imp, top = 10, radar = FALSE, xmeasure = "sumCover", ymeasure = "sumGain")

}
