% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceTableMixed.R
\name{importanceTableMixed}
\alias{importanceTableMixed}
\title{Importance Table Mixed}
\usage{
importanceTableMixed(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Importance Table Mixed
}
