% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsPlot.R
\name{interactionsPlot}
\alias{interactionsPlot}
\title{Interactions Plot}
\usage{
interactionsPlot(xgb.model, data, opt = "interactions")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{opt}{if "interactions" then strong interactions will be presented on the plot,
if "pairs" then plot presents all pairs in the model. Default "interactions".}
}
\value{
a ggplot object
}
\description{
Interactions Plot
}
\details{
NOTE: High gain of pair for \code{opt="pairs"} can be a result of high gain of down variable (child).
     As strong interactions should be considered only these pairs of variables,
     where variable on the bottom (child) has higher gain than variable on the top (parent).
}
\examples{

}
