% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsTable.R
\name{interactionsTable}
\alias{interactionsTable}
\title{Interactions table}
\usage{
interactionsTable(xgb.model, data, option = "interactions")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{option}{"interactions", "pairs"}
}
\value{
a data table
}
\description{
Interactions table
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

inter <- interactionsTable(xgb.model, sm,		option = "interactions")
inter
plot(inter)

inter <- interactionsTable(xgb.model, sm,		option = "pairs")
inter
plot(inter)

}
