% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopPlot.R
\name{lollipopPlot}
\alias{lollipopPlot}
\title{Visualiation of the model}
\usage{
lollipopPlot(xgb.model, data, labels = "topAll", log = TRUE)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{labels}{if "topAll" then labels for the most important interactions (vertical label)
and variables in the roots(horizontal label) will be displayed,
if "interactions" then labels for all interactions,
if "roots" then labels for all variables in the root.}

\item{log}{TRUE/FALSE logarithmic scale on the plot. Default TRUE.}
}
\value{
a ggplot object
}
\description{
Visualiation of the model
}
\examples{

}
