% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importanceTable.R
\name{plot.importanceTable}
\alias{plot.importanceTable}
\title{Importance plot}
\usage{
\method{plot}{importanceTable}(x, ..., top = 10, radar = TRUE,
  xmeasure = "sumCover", ymeasure = "sumGain")
}
\arguments{
\item{x}{a result of `importanceTable` function}

\item{...}{other parameters}

\item{top}{number of positions on the plot or NULL for all variable. Default 10.}

\item{radar}{TRUE/FALSE. If TRUE the plot shows
six measures of variables' importance in the model.
If FALSE the plot containing two chosen measures
of variables' importance in the model.}

\item{xmeasure}{measure on the x-axis.Available for `radar=FALSE`. Default "sumCover"}

\item{ymeasure}{measure on the y-axis. Available for `radar=FALSE`. Default "sumGain"}
}
\value{
a ggplot object
}
\description{
Importance plot
}
\details{
Available measures:
\itemize{
\item "sumGain" - sum of Gain value in all nodes, in which given variable occurs
\item "sumCover" - sum of Cover value in all nodes, in which given variable occurs; for LightGBM models: number of observation, which pass through the node
\item "mean5Gain" - mean gain from 5 occurrences of given variable with the highest gain
\item "meanGain" - mean Gain value in all nodes, in which given variable occurs
\item "meanCover" - mean Cover value in all nodes, in which given variable occurs; for LightGBM models: mean number of observation, which pass through the node
\item "freqency" - number of occurrences in the nodes for given variable
}

Additionally for plots with single variables:
\itemize{
\item "meanDepth"  - mean depth weighted by gain
\item "numberOfRoots" - number of occurrences in the root
\item "weightedRoot" - mean number of occurrences in the root, which is weighted by gain
}
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

imp <- importanceTable(xgb.model, sm, option = "both")
imp
plot(imp,  top = 10)

imp <- importanceTable(xgb.model, sm, option = "variables")
imp
plot(imp,  top = 10)

 imp <- importanceTable(xgb.model, sm, option = "interactions")
 imp
 plot(imp,  top = 10)

 imp <- importanceTable(xgb.model, sm, option = "variables")
 imp
 plot(imp, top = 10, radar = FALSE, xmeasure = "sumCover", ymeasure = "sumGain")

}
