% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactionsTable.R
\name{plot.interactionsTable}
\alias{plot.interactionsTable}
\title{Interactions plot}
\usage{
\method{plot}{interactionsTable}(x, ...)
}
\arguments{
\item{x}{a result of `interactionsTable` function}

\item{...}{other parameters}
}
\value{
a ggplot object
}
\description{
Interactions Plot
}
\details{
NOTE: High gain of pair for \code{option="pairs"} can be a result of high gain of down variable (child).
     As strong interactions should be considered only these pairs of variables,
     where variable on the bottom (child) has higher gain than variable on the top (parent).
}
\examples{
#' library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

inter <- interactionsTable(xgb.model, sm,		option = "interactions")
inter
plot(inter)

inter <- interactionsTable(xgb.model, sm,		option = "pairs")
inter
plot(inter)

}
