% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot.lollipop}
\alias{plot.lollipop}
\title{Visualiation of the model}
\usage{
\method{plot}{lollipop}(x, ..., labels = "topAll", log = TRUE)
}
\arguments{
\item{x}{a result of `EIX_lollipop` function}

\item{...}{other parameters}

\item{labels}{if "topAll" then labels for the most important interactions (vertical label)
and variables in the roots(horizontal label) will be displayed,
if "interactions" then labels for all interactions,
if "roots" then labels for all variables in the root.}

\item{log}{TRUE/FALSE logarithmic scale on the plot. Default TRUE.}
}
\value{
a ggplot object
}
\description{
Visualiation of the model
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

lolli <- EIX_lollipop(xgb.model, sm)
plot(lolli, labels = "topAll", log_scale = TRUE)

}
