% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarPlot.R
\name{radarPlot}
\alias{radarPlot}
\title{Importance radar plot}
\usage{
radarPlot(xgb.model, data, opt = "mixed", top = 10)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{opt}{if "single" then plot includes only single variable,
if "interactions", then only interactions
if "mixed", then both single variable and interactons.
Default "mixed".}

\item{top}{number of positions on the plot. Default 10.}
}
\value{
a ggplot object
}
\description{
The plot of six measures of variables' importance in the model.
}
\examples{

}
