% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootsGainTable.R
\name{rootsGainTable}
\alias{rootsGainTable}
\title{Function \code{rootsGainTable}}
\usage{
rootsGainTable(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{rootsGainTable}
}
