% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInteractionsPlot.R
\name{showInteractionsPlot}
\alias{showInteractionsPlot}
\title{Gain for interactions}
\usage{
showInteractionsPlot(xgb.model, data, opt = "interactions", trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{opt}{"interactions", "pairs"}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{showInteractionsPlot} returns plot with gains of each strong interaction in the model.
}
