% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPairsGainPlot.R
\name{showPairsGainPlot}
\alias{showPairsGainPlot}
\title{Gain plot for pairs}
\usage{
showPairsGainPlot(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{showPairsGainPlot} returns plot with gains of each pairs in the model.
}
\details{
NOTE Two variable are pairs when they is located in the model one below the other. If sum of gains is high, it does not mean, that there is strong interaction between this two variable.
To see the same plot for strong interactions see \code{showInterantionsGainPlot}.
}
