% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOfTrees.R
\name{tableOfTrees}
\alias{tableOfTrees}
\title{tableOfTrees}
\usage{
tableOfTrees(model, data)
}
\arguments{
\item{data}{a data table with data used to train the model}

\item{xgb.model}{a xgboost or lightgbm model}
}
\value{
a data table
}
\description{
tableOfTrees
}
