% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallPlot.R
\name{waterfallPlot}
\alias{waterfallPlot}
\title{waterfallPlot}
\usage{
waterfallPlot(xgb.model, new_observation, opt = "interactions")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{new_observation}{a new observation}

\item{opt}{"interactions", "single"}
}
\value{
a ggplot object
}
\description{
waterfallPlot
}
\examples{

}
