% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGain.R
\name{calculateGain}
\alias{calculateGain}
\title{calculateGain}
\usage{
calculateGain(xgb.model, data)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{
a list
}
\description{
List of trees with pairs of variable and other needed fields
}
\keyword{internal}
