% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPairs.R
\name{countPairs}
\alias{countPairs}
\title{Table with namber of pairs}
\usage{
countPairs(xgb.model, data, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{countPairs} counts how many times each pair of variable occures in the model.
}
