% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIX_lollipop.R
\name{EIX_lollipop}
\alias{EIX_lollipop}
\title{Tables needed for lollipop plot}
\usage{
EIX_lollipop(xgb.model, data)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{
an object of the lollipop class
}
\description{
Two tables needed to generate lollipop plot, which visualise the model.
The first table contains information about all nodes in the trees forming a model.
The second table includes information about roots in the trees.
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost(sm, params = param, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

lolli <- EIX_lollipop(xgb.model, sm)
plot(lolli, labels = "topAll", log_scale = TRUE)

}
