% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIX_waterfall.R
\name{EIX_waterfall}
\alias{EIX_waterfall}
\title{EIX_waterfall}
\usage{
EIX_waterfall(xgb.model, new_observation, option = "interactions",
  baseline = 0)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{new_observation}{a new observation}

\item{option}{"interaction" or "variables"}

\item{baseline}{number that specifies on which level the baseline line should be plotted}
}
\value{
an object of the broken class
}
\description{
Function \code{EIX_waterfall}
}
\details{
Function contains code or pieces of code
from `breakDown` code created by Przemysław Biecek
and`xgboostExplainer` code created by David Foster.
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost( param = param, data = sm, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

new_observation <- sm[9,]
wf <- EIX_waterfall(xgb.model, new_observation,  option = "interactions ")
wf
plot(wf)

}
