% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importacePlot.R, R/importancePlot.R
\name{importancePlot}
\alias{importancePlot}
\alias{importancePlot}
\title{Importance plot}
\usage{
importancePlot(xgb.model, data, xlab = "sumCover", ylab = "sumGain",
  opt = "mixed", top = 10, trees = NULL)

importancePlot(xgb.model, data, xlab = "sumCover", ylab = "sumGain",
  opt = "mixed", top = 10, trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{xlab}{measure on the x-axis. Default "sumCover"}

\item{ylab}{measure on the y-axis. Default "sumGain"}

\item{opt}{"single", "mixed","interactions". Default "mixed"}

\item{top}{number of variables on the plot or NULL for all variable. Default NULL}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}

\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{xlab}{measure on the x-axis. Default sumCover}

\item{ylab}{measure on the y-axis. Default SumGain}

\item{opt}{"single", "mixed","interactions". Default "mixed"}

\item{top}{number of variables on the plot or NULL for all variable. Default NULL}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Plot two chosen measures of variables importance in the xgboost model.

Function \code{importacePlot}
}
\details{
Avaible measures:
\itemize{
\item "sumGain"
\item "sumCover"
\item "mean5Gain"
\item "meanGain"
\item "meanCover"
\item "freqency"
}

Aditionally for plots with single variable:
\itemize{
\item "meanDepth",
\item "counterRoot",
\item "weightedRoot",
}
}
