% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceTable.R
\name{importanceTable}
\alias{importanceTable}
\title{Importance table}
\usage{
importanceTable(xgb.model, data, opt = "mixed")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{opt}{if "single" then table includes only single variable,
if "interactions", then only interactions
if "mixed", then both single variable and interactons.
Default "mixed".}
}
\value{
a data table
}
\description{
Importance table
}
\details{
Available measures:
\itemize{
\item "sumGain" - sum of Gain value in all nodes, in which given variable occurs
\item "sumCover" - sum of Cover value in all nodes, in which given variable occurs; for LightGBM models: number of observation, which pass through the node
\item "mean5Gain" - mean gain from 5 occurrences of given variable with the highest gain
\item "meanGain" - mean Gain value in all nodes, in which given variable occurs
\item "meanCover" - mean Cover value in all nodes, in which given variable occurs; for LightGBM models: mean number of observation, which pass through the node
\item "freqency" - number of occurrences in the nodes for given variable
}

Additionally for table with single variables:
\itemize{
\item "meanDepth"  - mean depth weighted by gain
\item "counterRoot" - number of occurrences in the root
\item "weightedRoot" - mean number of occurrences in the root, which is weighted by gain
}
}
\examples{

}
