% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceTable.R
\name{importanceTable}
\alias{importanceTable}
\title{Importance table}
\usage{
importanceTable(xgb.model, data, opt = "mixed", trees = NULL)
}
\arguments{
\item{xgb.model}{a xgboost model}

\item{data}{a DMatrix of data used to create the model}

\item{opt}{"single", "interactions","mixed". Default "mixed"}

\item{trees}{the number of trees to include in the xgboost model.Default NULL}
}
\description{
Function \code{importaceTable} returns table with importance of single variable and interactions with the following measure:
\itemize{
\item sumGain
\item sumCover
\item mean5Gain
\item meanGain
\item meanCover
\item freqency
}
}
