% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsTable.R
\name{interactionsTable}
\alias{interactionsTable}
\title{Importance of interactions and pairs in the model}
\usage{
interactionsTable(xgb.model, data, option = "interactions")
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}

\item{option}{if "interactions", the table contains interactions,
if "pairs", this table contains all the pairs in the model.
Default "interactions".}
}
\value{
a data table
}
\description{
The table with two measures of importance for interactions and pairs in the model.
}
\details{
Available measures:
\itemize{
\item "sumGain" - sum of Gain value in all nodes, in which given variable occurs
\item "freqency" - number of occurrences in the nodes for given variable
}

NOTE: High gain of pair for \code{option="pairs"} can be a result of high gain of down variable (child).
     As strong interactions should be considered only these pairs of variables,
     where variable on the bottom (child) has higher gain than variable on the top (parent).
}
\examples{
library("EIX")
library("Matrix")
library("data.table")
library("xgboost")

dt_HR <- data.table(HR_data)
sm <- sparse.model.matrix(left ~ . - 1,  data = dt_HR)

param <- list(objective = "binary:logistic", base_score = 0.5, max_depth = 2)
xgb.model <- xgboost(sm, params = param, label = dt_HR[, left] == 1, nrounds = 50, verbose = FALSE)

inter <- interactionsTable(xgb.model, sm, option = "interactions")
inter
plot(inter)

inter <- interactionsTable(xgb.model, sm, option = "pairs")
inter
plot(inter)

}
