% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGain.R
\name{tableOfTrees}
\alias{tableOfTrees}
\title{tableOfTrees}
\usage{
tableOfTrees(model, data)
}
\arguments{
\item{model}{a xgboost or lightgbm model}

\item{data}{a data table with data used to train the model}
}
\value{
a data table
}
\description{
tableOfTrees
}
\keyword{internal}
