% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Explain prediction of a single observation}
\usage{
waterfall(xgb.model, new_observation, option = "interactions", baseline = 0)
}
\arguments{
\item{xgb.model}{a xgboost or lightgbm model.}

\item{new_observation}{a new observation.}

\item{option}{if "variables", the plot includes only single variables,
if "interactions", then only interactions.
Default "interaction".}

\item{baseline}{a number or a character "Intercept" (for model intercept).
The baseline for the plot, where the rectangles should start.
Default 0.}
}
\value{
an object of the broken class
}
\description{
This function calculates a table with influence of variables and interactions
on the prediction of a given observation.
}
\details{
Function contains code or pieces of code
from \code{breakDown} code created by Przemysław Biecek
and \code{xgboostExplainer} code created by David Foster.
}
\examples{
library("EIX")
library("Matrix")
sm <- sparse.model.matrix(left ~ . - 1,  data = HR_data)

library("xgboost")
param <- list(objective = "binary:logistic", max_depth = 2)
xgb.model <- xgboost(sm, params = param, label = HR_data[, left] == 1, nrounds = 50, verbose=0)

new_observation <- sm[9,]
wf <- waterfall(xgb.model, new_observation,  option = "interactions")
wf
plot(wf)

}
